requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
numpy>=1.26.0
streamlit>=1.34.0
scikit-learn>=1.4.0
